requirements.txt
dj-database-url==0.4.1
Django==1.11.1
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-webpack-loader==0.2.4
djangorestframework==3.6.3
Babel==2.3.4
virtualenv==15.1.0
